feat: :art: Add carousel to projects section
Implement carousel functionality in the projects section  This commit introduces the integration of a carousel in the projects section of the Project component. A "keep-react" Carousel component was utilized to visually showcase projects in a carousel format.  Images from the projects array were used to populate the carousel slides, and options such as ring indicators and the display of three items at a time were configured. This significantly enhances the visual presentation of projects and allows users to navigate more efficiently between them.