feat: :zap: Hide sidebar
introduces changes to the style of the sidebar to hide it in the user interface. A "hidden" class has been added to the sidebar through the "sidebar.css" file. Clicking the toggle button will now change the visibility of the sidebar, providing a cleaner and focused user experience.  Local testing was conducted to ensure that the functionality of hiding the sidebar works correctly in various scenarios and devices.